refactor(7090): separate task lifecycle responsibilities
The lifecycle implementation is split into agreement, input, idempotency, and receipt modules so each state boundary remains reviewable and within repository size limits. Unreachable generic mutation helpers were removed after canonical routing replaced them.\n\nConstraint: Files remain at or below 200 lines and functions stay focused\nRejected: Suppress dead-code warnings | obsolete mutation paths would remain callable and misleading\nConfidence: high\nScope-risk: narrow\nDirective: Keep provider-bound reads side-effect free\nTested: Strict kamn-node all-target all-feature clippy; 30 task and escrow contracts\nNot-tested: Full repository make check pending integration phase
