+++ crates/kamn-node/src/main_tests/service_api_endpoint_tests/task_escrow_persistence_contract_tests/task_creator_provider_lifecycle_contract_tests.rs
@@ -1,10 +1,11 @@
 use super::super::*;
-use super::support::*;
-use crate::service_api_endpoint::{ServiceApiSnapshot, ServiceApiTaskTransitionBody};
-use std::path::PathBuf;
+use crate::service_api_endpoint::ServiceApiTaskTransitionBody;
 
+#[path = "task_creator_provider_lifecycle_contract_tests/support.rs"]
+mod lifecycle_support;
 #[path = "task_creator_provider_lifecycle_contract_tests/payloads.rs"]
 mod payloads;
+use lifecycle_support::LifecycleCase;
 use payloads::{assert_response, completion_body, retry_body, valid_create_body};
 
 const CREATOR: &str = "kamn:did:agent:task-lifecycle-creator";
@@ -120,7 +121,13 @@ fn integration_task_create_retry_is_idempotent_and_conflict_safe() {
         conflict_body.as_str(),
     );
     assert_response(&conflict, "409 Conflict", "TASK_IDEMPOTENCY_CONFLICT");
-    assert_eq!(case.state()["tasks"].as_object().unwrap().len(), 1);
+    assert_eq!(
+        case.state()["tasks"]
+            .as_object()
+            .expect("tasks should be an object")
+            .len(),
+        1
+    );
     case.cleanup();
 }
 
@@ -141,7 +148,9 @@ fn integration_task_transition_retry_returns_one_durable_receipt() {
         parse_service_api_payload(extract_http_response_body(&retried)).expect("retry payload");
     assert_eq!(retry_payload, first_payload);
     let state = case.state();
-    let receipts = state["task_transition_receipts"].as_array().unwrap();
+    let receipts = state["task_transition_receipts"]
+        .as_array()
+        .expect("transition receipts should be an array");
     assert_eq!(receipts.len(), 1);
     assert_eq!(receipts[0]["action"], "task:accept");
     assert_eq!(receipts[0]["transaction_id"], "transaction-lifecycle-001");
@@ -179,99 +188,11 @@ fn integration_task_completion_evidence_and_receipts_survive_restart() {
         "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"
     );
     assert_eq!(
-        state["task_transition_receipts"].as_array().unwrap().len(),
+        state["task_transition_receipts"]
+            .as_array()
+            .expect("transition receipts should be an array")
+            .len(),
         2
     );
     case.cleanup();
-}
-
-struct LifecycleCase {
-    _env: ServiceApiTestEnvGuards,
-    _state_guard: EnvVarGuard,
-    snapshot: ServiceApiSnapshot,
-    state_file: PathBuf,
-}
-
-impl LifecycleCase {
-    fn new(label: &str) -> Self {
-        let env = acquire_service_api_test_env();
-        let state_file = unique_named_state_file(format!("kamn-task-lifecycle-{label}").as_str());
-        let (_, state_guard) = set_state_file_env(state_file.as_path());
-        Self {
-            _env: env,
-            _state_guard: state_guard,
-            snapshot: build_task_escrow_snapshot("127.0.0.1:34220"),
-            state_file,
-        }
-    }
-
-    fn register(&self, actor: &str, nonce: u64) {
-        register_agent_profile(
-            &self.snapshot,
-            reserve_loopback_addr().as_str(),
-            actor,
-            nonce,
-            r#"{"agent_type":"provider","model_family":"pi","capabilities":["tasks"]}"#,
-        );
-    }
-
-    fn create_valid_task(&self, nonce: u64) -> ServiceApiTaskCreateBody {
-        let provider_did = test_service_api_sender_did(PROVIDER);
-        let body = valid_create_body(provider_did.as_str());
-        create_task(
-            &self.snapshot,
-            reserve_loopback_addr().as_str(),
-            CREATOR,
-            nonce,
-            body.as_str(),
-        )
-    }
-
-    fn authorized_request(
-        &self,
-        actor: &str,
-        nonce: u64,
-        method: &str,
-        path: &str,
-        body: &str,
-    ) -> String {
-        signed_request(&self.snapshot, actor, nonce, method, path, body, true)
-    }
-
-    fn raw_request(&self, actor: &str, nonce: u64, method: &str, path: &str, body: &str) -> String {
-        signed_request(&self.snapshot, actor, nonce, method, path, body, false)
-    }
-
-    fn state(&self) -> serde_json::Value {
-        read_state_json(self.state_file.as_path())
-    }
-
-    fn cleanup(self) {
-        let _ = fs::remove_file(self.state_file);
-    }
-}
-
-fn signed_request(
-    snapshot: &ServiceApiSnapshot,
-    actor: &str,
-    nonce: u64,
-    method: &str,
-    path: &str,
-    body: &str,
-    provision: bool,
-) -> String {
-    let request = SignedRequest {
-        max_requests: 1,
-        method,
-        path,
-        caller_did: actor,
-        nonce,
-        body,
-        extra_headers: &[("X-KAMN-Authz-Scope", "tasks:write")],
-    };
-    let bind_addr = reserve_loopback_addr();
-    if provision {
-        return authorized_signed_request(snapshot, bind_addr.as_str(), request);
-    }
-    raw_signed_request(snapshot, bind_addr.as_str(), request)
 }
