+++ crates/kamn-node/src/main_tests/service_api_endpoint_tests/task_escrow_persistence_contract_tests/task_escrow_routes_contract_tests/task_dispatch_support/payloads.rs
@@ -1,0 +1,11 @@
+pub(super) fn canonical_dispatch_body(provider: &str, task_type: &str, key: &str) -> String {
+    serde_json::json!({
+        "provider_did": provider,
+        "transaction_id": format!("transaction-{key}"),
+        "terms_digest": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+        "idempotency_key": key,
+        "task_type": task_type,
+        "description": "canonical dispatch task",
+    })
+    .to_string()
+}
